+++ backend/src/main/java/com/woowacourse/ternoko/support/alarm/AlarmSenderPicker.java
@@ -1,4 +1,4 @@
-package com.woowacourse.ternoko.support;
+package com.woowacourse.ternoko.support.alarm;
 
 public class AlarmSenderPicker {
