+++ backend/src/main/java/com/woowacourse/ternoko/support/alarm/SlackMessageGenerator.java
@@ -1,4 +1,4 @@
-package com.woowacourse.ternoko.support;
+package com.woowacourse.ternoko.support.alarm;
 
 import com.slack.api.methods.request.chat.ChatPostMessageRequest;
 import com.slack.api.model.Attachment;
